+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityModel.java
@@ -20,14 +20,14 @@ package is.codion.swing.framework.model;
 
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityType;
-import is.codion.framework.model.AbstractEntityModel;
+import is.codion.framework.model.DefaultEntityModel;
 
 import static java.util.Objects.requireNonNull;
 
 /**
  * A Swing implementation of {@link is.codion.framework.model.EntityModel}
  */
-public class SwingEntityModel extends AbstractEntityModel<SwingEntityModel,
+public class SwingEntityModel extends DefaultEntityModel<SwingEntityModel,
 				SwingEntityEditModel, SwingEntityTableModel, SwingEntityEditor> {
 
 	/**